complete 95%

+++ app/Http/Controllers/StudentController.php
@@ -25,51 +25,59 @@ class StudentController extends Controller
 
     public function store(Request $request)
     {
+        if (Gate::allows('isAdmin')) {
 
-        // validasi
-        $attr = $this->requestValidate();
+            // validasi
+            $attr = $this->requestValidate();
 
-        $user = new \App\User;
-        $user->role = 'student';
-        $user->name = $attr['nama_depan'];
-        $user->username = strtolower($attr['nama_depan']);
-        $user->email = $attr['email'];
-        $user->password = bcrypt('rahasia');
-        $user->remember_token = Str::random(60);
-        $user->save();
+            $user = new \App\User;
+            $user->role = 'student';
+            $user->name = $attr['nama_depan'];
+            $user->username = strtolower($attr['nama_depan']);
+            $user->email = $attr['email'];
+            $user->password = bcrypt('rahasia');
+            $user->remember_token = Str::random(60);
+            $user->save();
 
-        Student::create([
-            'user_id' => $user->id,
-            'nama_depan' => $request->nama_depan,
-            'nama_belakang' => $request->nama_belakang,
-            'jenis_kelamin' => $request->jenis_kelamin,
-            'pekerjaan' => $request->pekerjaan,
-            'kelas_paket' => $request->kelas_paket,
-            'alamat' => $request->alamat
-        ]);
-        //pesan flash message
-        session()->flash('success', 'Siswa baru telah dibuat');
+            Student::create([
+                'user_id' => $user->id,
+                'nama_depan' => $request->nama_depan,
+                'nama_belakang' => $request->nama_belakang,
+                'jenis_kelamin' => $request->jenis_kelamin,
+                'pekerjaan' => $request->pekerjaan,
+                'kelas_paket' => $request->kelas_paket,
+                'alamat' => $request->alamat
+            ]);
+            //pesan flash message
+            session()->flash('success', 'Siswa baru telah dibuat');
 
-        return redirect()->to('/admin/students');
+            return redirect()->to('/admin/students');
+        } else {
+            return abort(404);
+        }
     }
 
     // murid untuk update
     public function updatestudent(Student $student)
     {
-        $attr = $this->requestValidate();
-        if (request()->file('avatar')) {
-            \Storage::delete($student->avatar);
-            $avatar = request()->file('avatar')->store("images/siswa");
+        if (Gate::any(['isStudent', 'isAdmin'])) {
+            $attr = $this->requestValidate();
+            if (request()->file('avatar')) {
+                \Storage::delete($student->avatar);
+                $avatar = request()->file('avatar')->store("images/siswa");
+            } else {
+                $avatar = $student->avatar;
+            }
+            $user = User::where('id', $student->user_id)->first();
+            $attr['avatar'] = $avatar;
+            $user['email'] = $attr['email'];
+            $student->update($attr);
+            $user->update($attr);
+            session()->flash('success', 'Siswa Telah terupdate');
+            return redirect()->back();
         } else {
-            $avatar = $student->avatar;
+            return abort(404);
         }
-        $user = User::where('id', $student->user_id)->first();
-        $attr['avatar'] = $avatar;
-        $user['email'] = $attr['email'];
-        $student->update($attr);
-        $user->update($attr);
-        session()->flash('success', 'Siswa Telah terupdate');
-        return redirect()->back();
     }
 
     // public function destroy($idstudent)
@@ -82,21 +90,29 @@ class StudentController extends Controller
 
     public function addnilai(Request $request, $idstudent)
     {
-        $student = Student::find($idstudent);
-        if ($student->courses()->where('course_id', $request->courses)->exists()) {
-            session()->flash('delete', 'Nilai Sudah ada');
+        if (Gate::any(['isTeacher', 'isAdmin'])) {
+            $student = Student::find($idstudent);
+            if ($student->courses()->where('course_id', $request->courses)->exists()) {
+                session()->flash('delete', 'Nilai Sudah ada');
+                return redirect()->back();
+            }
+            $student->courses()->attach($request->courses, ['nilai' => $request->nilai]);
+            session()->flash('success', 'Nilai Telah masuk');
             return redirect()->back();
+        } else {
+            return abort(404);
         }
-        $student->courses()->attach($request->courses, ['nilai' => $request->nilai]);
-        session()->flash('success', 'Nilai Telah masuk');
-        return redirect()->back();
     }
 
     public function deletenilai(Student $student, $idcourse)
     {
-        $student->courses()->detach($idcourse);
-        session()->flash('delete', 'Mata Pelajaran Telah Di hapus');
-        return redirect()->back();
+        if (Gate::any(['isTeacher', 'isAdmin'])) {
+            $student->courses()->detach($idcourse);
+            session()->flash('delete', 'Mata Pelajaran Telah Di hapus');
+            return redirect()->back();
+        } else {
+            return abort(404);
+        }
     }
 
     public function requestValidate()
